requirements.txt
atlassian_python_api==3.41.19